backend/pinecone/requirements.txt
openai==0.27.4
openai[embeddings]
python-dotenv==1.0.0
nltk==3.8.1
tqdm==4.65.0
pinecone-client==3.0.1
fastapi==0.95.1
uvicorn==0.22.0
